requirements.txt
asgiref==3.2.3
DateTime==4.3
Django==3.0.3
django-auth-ldap==2.1.0
django-cors-headers==3.2.1
django-crontab==0.7.1
django-fontawesome-5==1.0.18
django-webpack-loader==0.7.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
mysqlclient==1.4.6
npm==0.1.1
numpy==1.18.1
optional-django==0.1.0
pandas==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
pyldap==3.0.0.post1
python-dateutil==2.8.1
python-ldap==3.2.0
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
zope.interface==4.7.1
